Moved all dbUtil-stuff together and finished up osmUtil clean-up
