+++ generate-osm/go.mod
@@ -3,10 +3,13 @@ module transform-osm
 go 1.19
 
 require (
+	github.com/urfave/cli/v2 v2.24.4 // direct
+	golang.org/x/net v0.7.0 // direct
+)
+
+require (
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
-	github.com/urfave/cli/v2 v2.24.1 // direct
-	golang.org/x/net v0.7.0 // direct
 	golang.org/x/text v0.7.0 // indirect
 )
